adjusted models order and item, showm all data
